% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SAS_output.R
\name{plot_SAS_output}
\alias{plot_SAS_output}
\title{standardized SAS output plots}
\usage{
plot_SAS_output(model_output, save = FALSE, dir_path = NULL)
}
\arguments{
\item{model_output}{output list produced by running run_SAS()}

\item{save}{a logical of whether or not you want the plots to be saved to disk}

\item{dir_path}{path to the directory where you wan the plots to be saved. Used only if save = TRUE.}
}
\value{
Returns four plots: Age quantiles, CDFs, Concentrations, PDFs, and Young Water Fractions.
}
\description{
Plots standardized output for trans-SAS soil models.
}
