% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SAS.R
\name{run_SAS}
\alias{run_SAS}
\title{run a SAS model}
\usage{
run_SAS(model_parameters = example_1_params)
}
\arguments{
\item{model_parameters}{a list of the model parameters. See Details.}
}
\value{
returns a list with model_parameters and model_output
}
\description{
run a SAS model
}
\examples{
data("example_1_params")
ex1model <- run_SAS(example_1_params)
#plot_SAS_input(ex1model$model_parameters)
#plot_SAS_output(ex1model$model_output)

}
\seealso{
\link{example_1_params}, \code{\link[=plot_SAS_input]{plot_SAS_input()}}, and \code{\link[=plot_SAS_output]{plot_SAS_output()}}
}
