% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS_shapes.R
\name{shape_beta}
\alias{shape_beta}
\title{beta distribution SAS shape}
\usage{
shape_beta(P_s, parameters, ...)
}
\arguments{
\item{P_s}{the cumulative storage age distribution (S_T/S)}

\item{parameters}{list of shape parameters for the beta distribution: 'a' and 'b'. Lower values of parameter 'a'
indicate a higher affinity for younger ages, whereas lower values of parameter 'b' indicate a higher affinity for
older ages.}

\item{...}{unused arguments}
}
\value{
returns the cumulative SAS function, Omega.
}
\description{
beta distribution SAS shape
}
