% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS_shapes.R
\name{shape_inv_pl}
\alias{shape_inv_pl}
\title{inverse power law SAS shape}
\usage{
shape_inv_pl(P_s, parameters, wetness = NULL)
}
\arguments{
\item{P_s}{the cumulative storage age distribution (S_T/S)}

\item{parameters}{A list with either parameter 'a' for a fixed exponent or parameters
'a_min' and 'a_max' corresponding to wetness index of 1 ('a_min') and wetness index of 0 ('a_max')}

\item{wetness}{The wetness index of the system. A value between 0 and 1.}
}
\value{
Returns the cumulative SAS function, Omega.
}
\description{
Calculates an inverse power law aSAS function: Omega = (1-P_s)^a.
}
