% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS_shapes.R
\name{shape_norm_trunc}
\alias{shape_norm_trunc}
\title{normal (truncated) SAS shape}
\usage{
shape_norm_trunc(P_s, parameters, ...)
}
\arguments{
\item{P_s}{the cumulative storage age distribution (S_T/S)}

\item{parameters}{A list of parameters 'm' for the mean and 'sd' for the standard deviation of the normal distribution.}

\item{...}{unused arguments}
}
\value{
Returns the cumulative SAS function, Omega.
}
\description{
Calculates a truncated normal SAS function.
}
