% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS_shapes.R
\name{shape_pl}
\alias{shape_pl}
\title{power law SAS shape}
\usage{
shape_pl(P_s, parameters, ...)
}
\arguments{
\item{P_s}{the cumulative storage age distribution (S_T/S)}

\item{parameters}{A list including the power law exponent, 'k'.}

\item{...}{unused arguments}
}
\value{
Returns the cumulative SAS function, Omega.
}
\description{
Function to compute a power law fSAS function: Om = P_s^k.
}
