% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS_shapes.R
\name{shape_pl_tv}
\alias{shape_pl_tv}
\title{time-variant power law SAS shape}
\usage{
shape_pl_tv(P_s, parameters, wetness)
}
\arguments{
\item{P_s}{the cumulative storage age distribution (S_T/S)}

\item{parameters}{list of parameters: 'k_min' and 'k_max'. 'k_min' corresponds
to a wetness index of 1 and 'k_max' corresponds to a wetness index of 0.}

\item{wetness}{the wetness index of the system. A value between 0 and 1.}
}
\value{
Returns the cumulative SAS function, Omega.
}
\description{
time-variant power law SAS shape
}
