% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SAS_input.R
\name{plot_SAS_input}
\alias{plot_SAS_input}
\title{Plots soil SAS inputs}
\usage{
plot_SAS_input(model_parameters, save = FALSE, file_path = NULL)
}
\arguments{
\item{model_parameters}{a list of SAS model input parameters}

\item{save}{a logical of whether or not you want the plot to be saved to disk as a .png}

\item{file_path}{used only if save = TRUE. The file path (including plot name) to be saved}
}
\description{
Produces plots of the inputs to the SAS model
}
